+++ __init__.py
@@ -1,4 +1,4 @@
-from .python.config import McuType  # noqa: F401
+from .python.constants import McuType  # noqa: F401
 from .python.serial import PandaSerial  # noqa: F401
 from .python import (Panda, PandaDFU, # noqa: F401
                      BASEDIR, pack_can_buffer, unpack_can_buffer, calculate_checksum,
